requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27
django-cors-headers>=4.3
psycopg2-binary>=2.9
dj-database-url>=2.1
python-decouple>=3.8
gunicorn>=22.0
whitenoise>=6.6
celery>=5.4
redis>=5.0
